refactor: split geometry_support into a package
Resolves code_critique.md §1 (module size). The 1654-line
geometry_support.py is replaced by a geometry_support/ package of ten
modules, each < 500 lines, with no change to observable table/label
output and no change to the public import surface.

The five state-light helpers (construct_excluded_mask, formatted_column,
circle_coverage, prep_row, and the body-selection cluster) become
module-level free functions so they are directly unit-testable without a
SPICE-backed Record. Record keeps the orchestration methods. Latent
correctness bugs are moved verbatim, not fixed, so this is a pure
structural change.
